@@ -1,18 +1,42 @@
 package ru.skillbox.services;
 
-import ru.skillbox.dto.*;
-import ru.skillbox.mappers.EntityMapper;
-import ru.skillbox.model.GeneralBlogInfo;
-import ru.skillbox.model.Post;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
+import ru.skillbox.dto.AuthResponse;
+import ru.skillbox.dto.AuthorizeUserRequest;
+import ru.skillbox.dto.CalendarResponse;
+import ru.skillbox.dto.CaptchaResponse;
+import ru.skillbox.dto.CommentResponse;
+import ru.skillbox.dto.ModerationRequest;
+import ru.skillbox.dto.NewCommentRequest;
+import ru.skillbox.dto.PasswordResetRequest;
+import ru.skillbox.dto.PlainPostDto;
+import ru.skillbox.dto.PostModerationStatus;
+import ru.skillbox.dto.PostRequest;
+import ru.skillbox.dto.PostWithCommentsResponse;
+import ru.skillbox.dto.PostsResponse;
+import ru.skillbox.dto.RegisterUserRequest;
+import ru.skillbox.dto.ResultResponse;
+import ru.skillbox.dto.TagDto;
+import ru.skillbox.dto.TagsResponse;
+import ru.skillbox.enums.ModerationStatus;
+import ru.skillbox.enums.SortMode;
+import ru.skillbox.enums.Vote;
+import ru.skillbox.mappers.EntityMapper;
+import ru.skillbox.model.CaptchaCode;
+import ru.skillbox.model.GeneralBlogInfo;
+import ru.skillbox.model.Post;
+import ru.skillbox.model.PostComment;
+import ru.skillbox.model.User;
 
 @Service
 public class ResponseService {
@@ -23,19 +47,24 @@ public class ResponseService {
     private TagService tagService;
     @Autowired
     private EntityMapper entityMapper;
-
+    @Autowired
+    private AuthService authService;
+    @Autowired
+    private CaptchaCodeService captchaCodeService;
+    @Autowired
+    private VotesService votesService;
+    @Autowired
+    private CommentsService commentsService;
 
     public CalendarResponse getCalendarResponse(LocalDateTime year) {
-        List<Post> allPostList = postService.getAllPostsFromRepository(true);
+        List<Post> allPostList = postService.getAllPostsFromRepository(true, ModerationStatus.ACCEPTED);
         List<Post> postList = postService.searchByDate(allPostList, year, LocalDateTime.now());
-
         Map<String, Long> postsCountPerYear = postList.stream()
-                .collect(Collectors.groupingBy(p -> p.getTime().toString().split(" ")[0],
-                        Collectors.counting()));
+            .collect(Collectors.groupingBy(p -> p.getTime().toString().split(" ")[0],
+                                           Collectors.counting()));
         List<Integer> postYears = postList.stream()
-                .map(p -> p.getTime().getYear())
-                .collect(Collectors.toList());
-
+            .map(p -> p.getTime().getYear())
+            .collect(Collectors.toList());
         CalendarResponse calendarResponse = new CalendarResponse();
         calendarResponse.setYears(postYears);
         calendarResponse.setPosts(postsCountPerYear);
@@ -46,11 +75,11 @@ public class ResponseService {
     public GeneralBlogInfo getGeneralBlogInfo() {
 
         return new GeneralBlogInfo("DevPub",
-                "Рассказы разработчиков",
-                "+7 903 666-44-55",
-                "mail@mail.ru",
-                "Дмитрий Сергеев",
-                "2005");
+                                   "Рассказы разработчиков",
+                                   "+7 903 666-44-55",
+                                   "mail@mail.ru",
+                                   "Дмитрий Сергеев",
+                                   "2005");
     }
 
     public TagsResponse getTagsResponse(String tagQuery) {
@@ -60,13 +89,17 @@ public class ResponseService {
         return tagsResponse;
     }
 
-    public PostsResponse getPostsResponse(Integer offset,
-                                          Integer limit,
-                                          String sortMode,
-                                          String searchQuery,
-                                          LocalDateTime ldt,
-                                          String tag) {
-        List<Post> postList = postService.getAllPostsFromRepository(true);
+    public ResponseEntity<PostsResponse> getPostsResponse(Integer offset,
+                                                          Integer limit,
+                                                          SortMode sortMode,
+                                                          String searchQuery,
+                                                          LocalDateTime ldt,
+                                                          String tag,
+                                                          User user,
+                                                          ModerationStatus status,
+                                                          Boolean isActive) {
+        status = status == null ? ModerationStatus.ACCEPTED : status;
+        List<Post> postList = postService.getAllPostsFromRepository(isActive, status);
         if (searchQuery != null) {
             postList = postService.searchByQuery(postList, searchQuery);
         }
@@ -77,27 +110,30 @@ public class ResponseService {
         if (tag != null) {
             postList = postService.searchByTag(postList, tag);
         }
+        if (user != null) {
+            postList = postService.searchByUser(postList, user);
+        }
         List<PlainPostDto> plainPostDtoList = postList.stream()
-                .map(entityMapper::postToPlainPostDto)
-                .collect(Collectors.toList());
+            .map(entityMapper::postToPlainPostDto)
+            .collect(Collectors.toList());
         if (sortMode != null) {
             plainPostDtoList = sortPlainPostDtoListByMode(plainPostDtoList, sortMode);
         }
-        return formPostsResponse(offset, limit, plainPostDtoList);
+        return new ResponseEntity<>(formPostsResponse(offset, limit, plainPostDtoList), HttpStatus.OK);
     }
 
-    private List<PlainPostDto> sortPlainPostDtoListByMode(List<PlainPostDto> list, String mode) {
+    private List<PlainPostDto> sortPlainPostDtoListByMode(List<PlainPostDto> list, SortMode mode) {
         switch (mode) {
-            case "best":
+            case BEST:
                 list.sort(Comparator.comparing(PlainPostDto::getLikeCount).reversed());
                 break;
-            case "early":
+            case EARLY:
                 list.sort(Comparator.comparing(PlainPostDto::getTime));
                 break;
-            case "recent":
+            case RECENT:
                 list.sort(Comparator.comparing(PlainPostDto::getTime).reversed());
                 break;
-            case "popular":
+            case POPULAR:
                 list.sort(Comparator.comparing(PlainPostDto::getCommentCount).reversed());
                 break;
         }
@@ -118,8 +154,205 @@ public class ResponseService {
         return postsResponse;
     }
 
+    public ResponseEntity<PostWithCommentsResponse> getPostWithCommentsResponse(Integer id) {
+        PostWithCommentsResponse postWithCommentsResponse = postService.getPostWithCommentsById(id);
+        if (postWithCommentsResponse == null) {
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(postWithCommentsResponse, HttpStatus.OK);
+    }
 
-    public PostWithCommentsResponse getPostWithCommentsResponse(Integer id) {
-        return postService.getPostWithCommentsById(id);
+    public ResponseEntity<ResultResponse> moderate(ModerationRequest moderationRequest) {
+        Optional<User> userOptional = authService.getAuthorizedUser();
+        if (userOptional.isEmpty()) {
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+        User user = userOptional.get();
+        if (!user.getIsModerator()) {
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
+        final Optional<Post> postOptional = postService.getPostById(moderationRequest.getPostId());
+        if (postOptional.isEmpty()) {
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+        }
+        final Post post = postOptional.get();
+
+        final User postModerator = post.getModeratedBy();
+        if (postModerator != null && !postModerator.equals(user)) {
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
+
+        Post updatedPost = postService.updatePostModerationStatus(user, post, moderationRequest.getDecision());
+        ResultResponse result = new ResultResponse();
+        result.setResult(updatedPost != null);
+        return new ResponseEntity<>(result, HttpStatus.OK);
+    }
+
+    public ResponseEntity<AuthResponse> login(AuthorizeUserRequest user) {
+        AuthResponse response = new AuthResponse();
+        User userFromDB = authService.loginUser(user);
+        if (!authService.isValidPassword(user.getPassword(), userFromDB.getPassword())) {
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
+        response.setUser(entityMapper.getAuthorizedUserDTO(userFromDB));
+        response.setResult(true);
+
+        return new ResponseEntity<>(response, HttpStatus.OK);
+    }
+
+    public ResponseEntity<AuthResponse> registerUser(RegisterUserRequest request) {
+        AuthResponse response = new AuthResponse();
+        response.setUser(entityMapper.getAuthorizedUserDTO(authService.registerUser(request)));
+        response.setResult(true);
+
+        return new ResponseEntity<>(response, HttpStatus.OK);
+    }
+
+    public ResponseEntity<ResultResponse> resetUserPassword(PasswordResetRequest request) {
+        ResultResponse response = new ResultResponse();
+        response.setResult(authService.resetUserPassword(request));
+
+        return new ResponseEntity<>(response, HttpStatus.OK);
+    }
+
+    public ResponseEntity<ResultResponse> logout() {
+        ResultResponse response = new ResultResponse();
+        authService.logoutUser();
+        response.setResult(true);
+        return new ResponseEntity<>(response, HttpStatus.OK);
+    }
+
+    public ResponseEntity<AuthResponse> getAuthorizedUserResponse(){
+        if (authService.getAuthorizedUser().isEmpty()) {
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+        AuthResponse response = new AuthResponse();
+        Optional<User> authorizedUser = authService.getAuthorizedUser();
+        authorizedUser.ifPresent(user -> response.setUser(entityMapper.getAuthorizedUserDTO(user)));
+        response.setResult(authService.getAuthorizedUser().isPresent());
+        return new ResponseEntity<>(response, HttpStatus.OK);
+    }
+
+    public ResponseEntity<ResultResponse> restorePassword(String email) {
+        ResultResponse response = new ResultResponse();
+        response.setResult(authService.restoreUserPassword(email));
+        return new ResponseEntity<>(response, HttpStatus.OK);
+    }
+
+    public ResponseEntity<CaptchaResponse> getCaptchaResponse() {
+        CaptchaResponse response = new CaptchaResponse();
+        CaptchaCode captchaCode = captchaCodeService.getCaptcha();
+        response.setImage("data:image/png;base64, "
+                              .concat(CaptchaCodeService.generateBase64Image(captchaCode.getCode())));
+        response.setSecret(captchaCode.getSecretCode());
+        return new ResponseEntity<>(response, HttpStatus.OK);
+    }
+
+    public ResponseEntity<PostsResponse> getModeratedPosts(Integer offset, Integer limit, ModerationStatus status) {
+        Optional<User> userOptional = authService.getAuthorizedUser();
+        if (userOptional.isEmpty()) {
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+        User authorizedUser = userOptional.get();
+        if (!authorizedUser.getIsModerator()) {
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
+        return getPostsResponse(offset, limit, null, null, null,
+                                null, authorizedUser, status, true);
+    }
+
+
+    public ResponseEntity<ResultResponse> addNewPost(PostRequest request) {
+        Optional<User> userOptional = authService.getAuthorizedUser();
+        if (userOptional.isEmpty()) {
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+        User user = userOptional.get();
+        Post savedPost = postService.savePost(null, request, user);
+        ResultResponse result = new ResultResponse();
+        result.setResult(savedPost != null);
+        return new ResponseEntity<>(result, HttpStatus.OK);
+    }
+
+    public ResponseEntity<ResultResponse> editPost(Integer id, PostRequest request) {
+        Optional<User> userOptional = authService.getAuthorizedUser();
+        if (userOptional.isEmpty()) {
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+        User user = userOptional.get();
+        Optional<Post> postOptional = postService.getPostById(id);
+        if (postOptional.isEmpty()) {
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+        }
+        Post post = postOptional.get();
+        if (!post.getUser().equals(user) &&
+            (!user.getIsModerator() || (post.getModeratedBy() != null &&
+                !post.getModeratedBy().equals(user)))) {
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        }
+        Post savedPost = postService.savePost(post, request, user);
+        ResultResponse result = new ResultResponse();
+        result.setResult(savedPost != null);
+        return new ResponseEntity<>(result, HttpStatus.OK);
+    }
+
+    public ResponseEntity<PostsResponse> getMyPosts(Integer offset, Integer limit, PostModerationStatus status) {
+        Optional<User> userOptional = authService.getAuthorizedUser();
+        if (userOptional.isEmpty()) {
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+        User authorizedUser = userOptional.get();
+        status = status == null ? PostModerationStatus.PUBLISHED : status;
+        return getPostsResponse(offset, limit, null, null, null,
+                                null, authorizedUser, status.getModerationStatus(), status.isActive());
+    }
+
+    public ResponseEntity<ResultResponse> votePost(Vote vote, Integer postId) {
+        Optional<User> userOptional = authService.getAuthorizedUser();
+        if (userOptional.isEmpty()) {
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+        User user = userOptional.get();
+        if (postId <= 0) {
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+        }
+        Post post = postService.getPostById(postId).orElse(null);
+        if (post == null) {
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+        }
+
+        ResultResponse result = new ResultResponse();
+        result.setResult(votesService.vote(vote, userOptional.get(), post));
+        return new ResponseEntity<>(result, HttpStatus.OK);
+    }
+
+    public ResponseEntity<ResultResponse> comment(NewCommentRequest comment) {
+        Optional<User> userOptional = authService.getAuthorizedUser();
+        if (userOptional.isEmpty()) {
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+        }
+        User user = userOptional.get();
+        Optional<Post> post = postService.getPostById(comment.getPostId());
+        Optional<PostComment> parentComment = Optional.empty();
+        if (post.isEmpty()) {
+            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        }
+        if (comment.getParentId() != null) {
+            List<PostComment> postComments = post.get().getPostComments();
+            parentComment = commentsService.findCommentById(comment.getParentId());
+            if (parentComment.isEmpty() || (!postComments.isEmpty() && !postComments.contains(parentComment.get()))) {
+                return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+            }
+        }
+        CommentResponse result = new CommentResponse();
+        PostComment newComment = commentsService.addComment(user, parentComment.orElse(null),
+                                                            post.get(), comment.getText());
+        if (newComment != null) {
+            result.setResult(true);
+            result.setId(newComment.getId());
+        } else {
+            result.setResult(false);
+        }
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 }
