@@ -2,13 +2,13 @@ package ru.skillbox.controllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 
 @Controller
 public class DefaultController {
 
-    @RequestMapping("/")
-    public String index(){
-        return "index";
+    @RequestMapping(method = {RequestMethod.OPTIONS, RequestMethod.GET}, value = "/**/{path:[^\\.]*}")
+    public String redirectToIndex() {
+        return "forward:/";
     }
-
 }
