@@ -1,9 +1,14 @@
 package ru.skillbox.enums;
 
-public enum ModerationStatus {
-    NEW,
-    ACCEPTED,
-    DECLINED;
+import org.springframework.core.convert.converter.Converter;
 
-    ModerationStatus(){}
+public enum ModerationStatus {
+    NEW, ACCEPTED, DECLINED;
+
+    public static class StringToEnumConverter implements Converter<String, ModerationStatus> {
+        @Override
+        public ModerationStatus convert(String source) {
+            return ModerationStatus.valueOf(source.toUpperCase());
+        }
+    }
 }
