@@ -1,54 +1,33 @@
 package ru.skillbox.model;
 
-import javax.persistence.*;
-import javax.validation.constraints.NotNull;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import lombok.Data;
+import ru.skillbox.enums.GlobalSettings;
 
+@Data
 @Entity
-@Table(name="global_settings")
+@Table(name = "global_settings")
 public class GlobalSetting {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Integer id;
 
-    @NotNull
-    private String code;
+    @Enumerated(EnumType.STRING)
+    @Column(length = 30, nullable = false)
+    private GlobalSettings.Code code;
 
-    @NotNull
+    @Column(nullable = false)
     private String name;
 
-    @NotNull
-    private String value;
+    @Enumerated(EnumType.STRING)
+    @Column(length = 5, nullable = false)
+    private GlobalSettings.Value value;
 
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public void setCode(String code) {
-        this.code = code;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getValue() {
-        return value;
-    }
-
-    public void setValue(String value) {
-        this.value = value;
-    }
 }
