@@ -1,98 +1,47 @@
 package ru.skillbox.model;
 
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.ToString;
 import org.hibernate.annotations.Type;
 
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
+@Data
 @Entity
 @Table(name = "users")
+@EqualsAndHashCode(of = {"id"})
+@ToString(callSuper = true, of = {"name"})
 public class User {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Integer id;
 
     @Column(name = "is_moderator")
     @Type(type = "org.hibernate.type.NumericBooleanType")
-    private boolean isModerator;
+    private Boolean isModerator;
 
     @Column(name = "reg_time")
     private LocalDateTime regTime;
 
     private String name;
-
     private String email;
-
     private String password;
-
     private String code;
 
     @Column(length = 65535, columnDefinition="TEXT")
     @Type(type="text")
     private String photo;
 
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public boolean isModerator() {
-        return isModerator;
-    }
-
-    public void setModerator(boolean moderator) {
-        isModerator = moderator;
-    }
-
-    public LocalDateTime getRegTime() {
-        return regTime;
-    }
-
-    public void setRegTime(LocalDateTime regTime) {
-        this.regTime = regTime;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public void setCode(String code) {
-        this.code = code;
-    }
-
-    public String getPhoto() {
-        return photo;
-    }
-
-    public void setPhoto(String photo) {
-        this.photo = photo;
-    }
+    @OneToMany(mappedBy = "user", fetch = FetchType.LAZY, orphanRemoval = true)
+    private final List<PostVote> votes = new ArrayList<>();
 }
